requirements.txt
google-generativeai
playwright
pytesseract
pillow
pandas
openpyxl
thefuzz
python-dotenv